Extract from MainPage separated Notification component PageNotification.tsx
