+++ app/containers/MainPage.tsx
@@ -23,10 +23,6 @@ import { withStyles } from '@material-ui/core/styles';
 import { translate } from 'react-i18next';
 import SplitPane from 'react-split-pane';
 import SwipeableDrawer from '@material-ui/core/SwipeableDrawer';
-import Snackbar from '@material-ui/core/Snackbar';
-import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
-import CloseIcon from '@material-ui/icons/Close';
 import { HotKeys } from 'react-hotkeys';
 import { NativeTypes } from 'react-dnd-html5-backend';
 import { Progress } from 'aws-sdk/clients/s3';
@@ -43,12 +39,10 @@ import {
   getKeyBindingObject,
   getLeftVerticalSplitSize,
   getMainVerticalSplitSize,
-  getLastPublishedVersion,
   actions as SettingsActions
 } from '../reducers/settings';
 import {
   actions as AppActions,
-  getNotificationStatus,
   isGeneratingThumbs,
   isAboutDialogOpened,
   isOnboardingDialogOpened,
@@ -56,7 +50,6 @@ import {
   isLicenseDialogOpened,
   isThirdPartyLibsDialogOpened,
   isEntryInFullWidth,
-  isUpdateAvailable,
   getDirectoryPath,
   isLocationManagerPanelOpened,
   isTagLibraryPanelOpened,
@@ -77,17 +70,12 @@ import {
   getSelectedEntries,
   currentUser
 } from '../reducers/app';
-import {
-  actions as LocationIndexActions,
-  isIndexing
-} from '../reducers/location-index';
 import { buffer } from '-/utils/misc';
 import TargetFileBox from '../components/TargetFileBox';
 import AppConfig from '../config';
 import buildDesktopMenu from '../services/electron-menus';
 import buildTrayIconMenu from '../services/electron-tray-menu';
 import i18n from '../services/i18n';
-import { Pro } from '../pro';
 import LoadingLazy from '../components/LoadingLazy';
 import withDnDContext from '-/containers/withDnDContext';
 import { CustomDragLayer } from '-/components/CustomDragLayer';
@@ -97,7 +85,7 @@ import ProgressDialog from '-/components/dialogs/ProgressDialog';
 import useEventListener from '-/utils/useEventListener';
 import ConfirmDialog from '-/components/dialogs/ConfirmDialog';
 import { TS } from '-/tagspaces.namespace';
-import useTraceUpdate from '-/utils/useTraceUpdate';
+import PageNotification from '-/containers/PageNotification';
 
 const initialSplitSize = 44;
 const drawerWidth = 300;
@@ -144,15 +132,11 @@ interface Props {
   setFirstRun: (isFirstRun: boolean) => void;
   isDesktopMode: boolean;
   openedFiles: Array<OpenedEntry>;
-  isIndexing: boolean;
   isGeneratingThumbs: boolean;
   setGeneratingThumbnails: (isGenerating: boolean) => void;
   isEntryInFullWidth: boolean;
   classes: any;
   theme: any;
-  notificationStatus: any;
-  lastPublishedVersion: string;
-  isUpdateAvailable: boolean;
   isReadOnlyMode: boolean;
   isSettingsDialogOpened: boolean;
   isCreateFileDialogOpened: boolean;
@@ -173,10 +157,7 @@ interface Props {
   keyBindings: any;
   toggleEditTagDialog: () => void;
   setEntryFullWidth: (isFullWidth: boolean) => void;
-  hideNotifications: () => void;
-  cancelDirectoryIndexing: () => void;
   loadParentDirectoryContent: () => void;
-  setUpdateAvailable: (isUpdateAvailable: boolean) => void;
   saveFile: () => void; // needed by electron-menus
   setZoomResetApp: () => void; // needed by electron-menus
   setZoomInApp: () => void; // needed by electron-menus
@@ -377,19 +358,12 @@ const MainPage = (props: Props) => {
     return props.mainSplitSize;
   };
 
-  const isManagementPanelVisible = () => {
-    if (
-      props.isLocationManagerPanelOpened ||
-      props.isTagLibraryPanelOpened ||
-      props.isSearchPanelOpened ||
-      props.isPerspectivesPanelOpened ||
-      props.isHelpFeedbackPanelOpened
-    ) {
-      return true;
-    }
-    return false;
-    // return window.ExtDefaultVerticalPanel !== 'none' && !props.isEntryInFullWidth;
-  };
+  const isManagementPanelVisible = () =>
+    props.isLocationManagerPanelOpened ||
+    props.isTagLibraryPanelOpened ||
+    props.isSearchPanelOpened ||
+    props.isPerspectivesPanelOpened ||
+    props.isHelpFeedbackPanelOpened;
 
   useEventListener('resize', () => {
     if (!AppConfig.isCordova) {
@@ -442,27 +416,6 @@ const MainPage = (props: Props) => {
       props.openLocationManagerPanel();
     }
     // setDrawerOpened(true);
-  };
-
-  const skipRelease = () => {
-    props.setUpdateAvailable(false);
-  };
-
-  const getLatestVersion = () => {
-    if (Pro) {
-      props.showNotification(
-        i18n.t('core:getLatestVersionPro'),
-        'default',
-        false
-      );
-    } else {
-      props.openURLExternally(AppConfig.links.downloadURL, true);
-    }
-    props.setUpdateAvailable(false);
-  };
-
-  const openChangelogPage = () => {
-    props.openURLExternally(AppConfig.links.changelogURL, true);
   };
 
   const handleFileDrop = (item, monitor) => {
@@ -668,90 +621,7 @@ const MainPage = (props: Props) => {
           confirmDialogContentTID="confirmDeleteDialogContent"
         />
       )}
-      <Snackbar
-        data-tid={props.notificationStatus.tid}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        open={props.notificationStatus.visible}
-        onClose={props.hideNotifications}
-        autoHideDuration={props.notificationStatus.autohide ? 3000 : undefined}
-        message={props.notificationStatus.text}
-        action={[
-          <IconButton
-            data-tid={'close' + props.notificationStatus.tid}
-            key="close"
-            aria-label={i18n.t('core:closeButton')}
-            color="inherit"
-            onClick={props.hideNotifications}
-          >
-            <CloseIcon />
-          </IconButton>
-        ]}
-      />
-      {/* <Snackbar
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        open={props.isGeneratingThumbs}
-        autoHideDuration={undefined}
-        message={i18n.t('core:loadingOrGeneratingThumbnails')}
-        action={[
-          <IconButton
-            key="closeButton"
-            aria-label={i18n.t('core:closeButton')}
-            color="inherit"
-            onClick={() => props.setGeneratingThumbnails(false)}
-          >
-            <CloseIcon />
-          </IconButton>
-        ]}
-      /> */}
-      <Snackbar
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        open={props.isIndexing}
-        autoHideDuration={undefined}
-        message="Indexing"
-        action={[
-          <Button
-            key="cancelIndexButton"
-            color="secondary"
-            size="small"
-            onClick={props.cancelDirectoryIndexing}
-            data-tid="cancelDirectoryIndexing"
-          >
-            {i18n.t('core:cancelIndexing')}
-          </Button>
-        ]}
-      />
-      <Snackbar
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-        open={props.isUpdateAvailable}
-        autoHideDuration={undefined}
-        message={'Version ' + props.lastPublishedVersion + ' available.'}
-        action={[
-          <Button
-            key="laterButton"
-            color="secondary"
-            size="small"
-            onClick={skipRelease}
-          >
-            {i18n.t('core:later')}
-          </Button>,
-          <Button
-            key="changelogButton"
-            color="secondary"
-            size="small"
-            onClick={openChangelogPage}
-          >
-            {i18n.t('core:releaseNotes')}
-          </Button>,
-          <Button
-            key="latestVersionButton"
-            color="primary"
-            size="small"
-            onClick={getLatestVersion}
-          >
-            {i18n.t('core:getItNow')}
-          </Button>
-        ]}
-      />
+      <PageNotification />
       {props.isDesktopMode || (AppConfig.isAmplify && !props.user) ? (
         <TargetFileBox
           // @ts-ignore
@@ -861,7 +731,6 @@ function mapStateToProps(state) {
     isUploadProgressDialogOpened: isUploadDialogOpened(state),
     isOpenLinkDialogOpened: isOpenLinkDialogOpened(state),
     isProgressDialogOpened: isProgressOpened(state),
-    isIndexing: isIndexing(state),
     isReadOnlyMode: isReadOnlyMode(state),
     isGeneratingThumbs: isGeneratingThumbs(state),
     // isFileOpened: isFileOpened(state),
@@ -871,9 +740,6 @@ function mapStateToProps(state) {
     keyBindings: getKeyBindingObject(state),
     leftSplitSize: getLeftVerticalSplitSize(state),
     mainSplitSize: getMainVerticalSplitSize(state),
-    isUpdateAvailable: isUpdateAvailable(state),
-    lastPublishedVersion: getLastPublishedVersion(state),
-    notificationStatus: getNotificationStatus(state),
     isLocationManagerPanelOpened: isLocationManagerPanelOpened(state),
     isTagLibraryPanelOpened: isTagLibraryPanelOpened(state),
     isSearchPanelOpened: isSearchPanelOpened(state),
@@ -897,9 +763,7 @@ function mapDispatchToProps(dispatch) {
       toggleProgressDialog: AppActions.toggleProgressDialog,
       resetProgress: AppActions.resetProgress,
       toggleEditTagDialog: AppActions.toggleEditTagDialog,
-      hideNotifications: AppActions.hideNotifications,
       onUploadProgress: AppActions.onUploadProgress,
-      cancelDirectoryIndexing: LocationIndexActions.cancelDirectoryIndexing,
       saveFile: AppActions.saveFile,
       setZoomResetApp: SettingsActions.setZoomResetApp,
       setZoomInApp: SettingsActions.setZoomInApp,
@@ -919,7 +783,6 @@ function mapDispatchToProps(dispatch) {
       openFsEntry: AppActions.openFsEntry,
       openURLExternally: AppActions.openURLExternally,
       setEntryFullWidth: AppActions.setEntryFullWidth,
-      setUpdateAvailable: AppActions.setUpdateAvailable,
       openNextFile: AppActions.openNextFile,
       openPrevFile: AppActions.openPrevFile,
       toggleShowUnixHiddenEntries: SettingsActions.toggleShowUnixHiddenEntries,
@@ -955,7 +818,6 @@ const areEqual = (prevProp, nextProp) =>
   nextProp.isEditTagDialogOpened === prevProp.isEditTagDialogOpened &&
   nextProp.isEntryInFullWidth === prevProp.isEntryInFullWidth &&
   nextProp.isHelpFeedbackPanelOpened === prevProp.isHelpFeedbackPanelOpened &&
-  nextProp.isIndexing === prevProp.isIndexing &&
   nextProp.isKeysDialogOpened === prevProp.isKeysDialogOpened &&
   nextProp.isLicenseDialogOpened === prevProp.isLicenseDialogOpened &&
   nextProp.isLocationManagerPanelOpened ===
@@ -970,13 +832,10 @@ const areEqual = (prevProp, nextProp) =>
   nextProp.isTagLibraryPanelOpened === prevProp.isTagLibraryPanelOpened &&
   nextProp.isThirdPartyLibsDialogOpened ===
     prevProp.isThirdPartyLibsDialogOpened &&
-  nextProp.isUpdateAvailable === prevProp.isUpdateAvailable &&
   nextProp.isUploadProgressDialogOpened ===
     prevProp.isUploadProgressDialogOpened &&
   nextProp.leftSplitSize === prevProp.leftSplitSize &&
   nextProp.mainSplitSize === prevProp.mainSplitSize &&
-  JSON.stringify(nextProp.notificationStatus) ===
-    JSON.stringify(prevProp.notificationStatus) &&
   JSON.stringify(nextProp.selectedEntries) ===
     JSON.stringify(prevProp.selectedEntries) &&
   JSON.stringify(nextProp.openedFiles) === JSON.stringify(prevProp.openedFiles);
